debug

+++ samples/pairing/c_pre.c
@@ -770,13 +770,13 @@ int Enc2(unsigned char *pk_Hex_data, int pk_Hex_data_len)
     unsigned char *pk_data[G1_ELEMENT_LENGTH_IN_BYTES];
     printf("pk_Hex_data=\n");
     for(int i=0;i<G1_ELEMENT_LENGTH_IN_BYTES * 2;i++) {
-        printf("%c", (unsigned int)pk_Hex_data[i]);
+        printf("%c", pk_Hex_data[i]);
     }
     printf("\n");
     int iret = HexStrToByteStr((uint8_t *)pk_Hex_data, pk_Hex_data_len, (uint8_t *)pk_data);
     printf("pk_data=\n");
     for(int i=0;i<G1_ELEMENT_LENGTH_IN_BYTES;i++) {
-        printf("%02x ", (unsigned int)pk_data[i]);
+        printf("%02x ", pk_data[i]);
     }
     printf("\n");
